+++ internal/ngsilib/token_password_credentials.go
@@ -30,9 +30,11 @@ SOFTWARE.
 package ngsilib
 
 import (
+	"encoding/base64"
 	"fmt"
 	"net/http"
 	"net/url"
+	"strings"
 	"time"
 )
 
@@ -80,6 +82,35 @@ func (i *idmPasswordCredentials) requestToken(ngsi *NGSI, client *Client, tokenI
 	tokenInfo.Oauth = &token
 
 	return tokenInfo, nil
+}
+
+func (i *idmPasswordCredentials) revokeToken(ngsi *NGSI, client *Client, tokenInfo *TokenInfo) error {
+	const funcName = "revokeTokenPasswordCredentials"
+
+	headers := make(map[string]string)
+
+	s := strings.Replace(client.idmURL(), "/token", "/revoke", 1)
+	if !strings.HasSuffix(s, "/revoke") {
+		s += "/revoke"
+	}
+	u, _ := url.Parse(s)
+	idm := Client{URL: u, Headers: headers, HTTP: ngsi.HTTP}
+	broker := client.Server
+
+	idm.SetHeader(cContentType, cAppXWwwFormUrlencoded)
+	auth := fmt.Sprintf("%s:%s", broker.ClientID, broker.ClientSecret)
+	idm.SetHeader("Authorization", fmt.Sprintf("Basic %s", base64.StdEncoding.EncodeToString([]byte(auth))))
+	payload := fmt.Sprintf("token=%s&token_type_hint=refresh_token", tokenInfo.RefreshToken)
+
+	res, body, err := idm.HTTPPost(payload)
+	if err != nil {
+		return &LibError{funcName, 1, err.Error(), err}
+	}
+	if res.StatusCode != http.StatusOK {
+		return &LibError{funcName, 2, fmt.Sprintf("error %s %s", res.Status, string(body)), nil}
+	}
+
+	return nil
 }
 
 func (i *idmPasswordCredentials) getAuthHeader(token string) (string, string) {